refactor(sidecar): api spec module
