+++ bolt-sidecar/src/api/mod.rs
@@ -1,1 +1,2 @@
 pub mod builder;
+pub mod spec;
